closer-to-working bytecode change to int

+++ scrub-lang/Compiler/CompilationScope.cs
@@ -2,7 +2,7 @@
 
 public class CompilationScope
 {
-	public List<byte> Instuctions = new List<byte>();
+	public List<int> Instuctions = new List<int>();
 	public EmittedInstruction LastInstruction;//last instruction
 	public EmittedInstruction PreviousInstruction;//the one before lastInstructin.
 	//public Dictionary<int, Location> LocationTable = new Dictionary<int, Location>(); //connects instucton
